move RocketmqProperties to properties
