+++ eventmesh-dashboard-core/src/main/java/org/apache/eventmesh/dashboard/core/store/RocketmqTopicCore.java
@@ -18,8 +18,8 @@
 package org.apache.eventmesh.dashboard.core.store;
 
 
-import org.apache.eventmesh.dashboard.common.model.RocketmqProperties;
-import org.apache.eventmesh.dashboard.common.model.TopicProperties;
+import org.apache.eventmesh.dashboard.common.properties.RocketmqProperties;
+import org.apache.eventmesh.dashboard.common.dto.TopicProperties;
 import org.apache.eventmesh.dashboard.common.util.RocketmqUtils;
 import org.apache.eventmesh.dashboard.service.store.TopicCore;
